requirements.txt
aiosmtplib==1.1.6
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
Django==5.2.3
djangorestframework==3.16.0
drf-spectacular==0.28.0
et-xmlfile==1.1.0
fake-headers==1.0.2
filelock==3.13.1
h11==0.14.0
html5lib==1.1
idna==3.6
inflection==0.5.1
Jinja2==3.0.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
lxml==5.1.0
MarkupSafe==2.1.3
numpy==1.26.1
openpyxl==3.0.7
outcome==1.3.0.post0
packaging==23.2
pdfkit-async==0.7.2
platformdirs==4.1.0
pycparser==2.21
PySocks==1.7.1
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.31.0
rpds-py==0.26.0
selenium==4.18.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.2.0
urllib3==2.2.1
uvicorn==0.15.0
virtualenv==20.25.0
webdriver-manager==4.0.1
webencodings==0.5.1
wsproto==1.2.0
